debundle: share owner-graph gate fixtures via support helpers
Three byte-identical fixtures were copy-pasted across the gate CLI tests:

- graph_with_atomic_unit (bindings_assign_gate, bindings_unassign_gate)
- write_atomic_unit_fixture (bindings_assign_gate, bindings_unassign_gate)
- graph_with_acyclic_cross_module_read (bindings_assign_gate, modules_gate)

Hoist all three into debundle_e2e_support. They are plain owner-graph JSON
strings (no :analysis dependency), so they live in :support like the other
fixture builders. owner_for_binding stays local: it needs OwnerGraphReport
from :analysis, which the standard e2e tests deliberately keep out of their
compile graph.

Net -97 LOC.

BAZEL_TEST_INVOCATIONS=buildbuddy:34c79b2e-8a07-4788-a6c0-faa361ff70dd
